+++ src/Products/PloneMeeting/tests/testAnnexes.py
@@ -26,7 +26,7 @@ from AccessControl import Unauthorized
 from collective.contact.plonegroup.utils import get_plone_group
 from collective.documentviewer.config import CONVERTABLE_TYPES
 from collective.documentviewer.settings import GlobalSettings
-from collective.iconifiedcategory.event import IconifiedPrintChangedEvent
+from collective.iconifiedcategory.event import IconifiedAttrChangedEvent
 from collective.iconifiedcategory.interfaces import IIconifiedPreview
 from collective.iconifiedcategory.utils import calculate_category_id
 from collective.iconifiedcategory.utils import get_categorized_elements
@@ -831,18 +831,20 @@ class testAnnexes(PloneMeetingTestCase):
         self.assertFalse(converted_annex.to_print)
         self.assertFalse(IIconifiedPreview(converted_annex).converted)
         converted_annex.to_print = True
-        notify(IconifiedPrintChangedEvent(converted_annex,
-                                          old_values={'to_print': False},
-                                          new_values={'to_print': True}))
+        notify(IconifiedAttrChangedEvent(converted_annex,
+                                         attr_name='to_print',
+                                         old_values={'to_print': False},
+                                         new_values={'to_print': True}))
         self.assertTrue(converted_annex.to_print)
         self.assertTrue(IIconifiedPreview(converted_annex).converted)
 
         # if an annex is not 'to_print', it is not converted
         converted_annex2 = self.addAnnex(item)
         converted_annex2.to_print = False
-        notify(IconifiedPrintChangedEvent(converted_annex2,
-                                          old_values={'to_print': True},
-                                          new_values={'to_print': False}))
+        notify(IconifiedAttrChangedEvent(converted_annex2,
+                                         attr_name='to_print',
+                                         old_values={'to_print': True},
+                                         new_values={'to_print': False}))
         self.assertFalse(converted_annex2.to_print)
         self.assertFalse(IIconifiedPreview(converted_annex2).converted)